Fix encoding breaking sometimes
The issue wasn't with sys.maxsize or the value it was being multiplied by.

When a string is being encrypted, the values of the multiplier are hidden within the encrypted text. The positions of where these values are are randomly generated, however, it was never checked that there would be duplicate values, leading to some positions being overwritten and causing problems.